[CARBONDATA-2924] Fix parsing issue for map as a nested array child and change the error message in sort column validation for SDK
This PR contains

Fix to support nested complex type till integer length. When the data length in one row for nested complex type was crossing the short limit exception was being throw. Now length till integer limit is supported.
Fixed error message when map type was given as sort column in SDK
Fixed map type parsing issue for nested structures like struct<array> or array<array>
Modified code to add to LRU cache only when the flag to add to cache is set to true

This closes #2702

+++ processing/src/main/java/org/apache/carbondata/processing/loading/sort/SortStepRowHandler.java
@@ -267,7 +267,7 @@ public class SortStepRowHandler implements Serializable {
 
     // read complex dims
     for (int i = 0; i < complexDimCnt; i++) {
-      short len = rowBuffer.getShort();
+      int len = rowBuffer.getInt();
       byte[] bytes = new byte[len];
       rowBuffer.get(bytes);
       noDictDims[noDictIndex++] = bytes;
@@ -587,7 +587,7 @@ public class SortStepRowHandler implements Serializable {
     // convert complex dims
     for (int idx = 0; idx < this.complexDimCnt; idx++) {
       byte[] bytes = (byte[]) row[this.complexDimIdx[idx]];
-      rowBuffer.putShort((short) bytes.length);
+      rowBuffer.putInt(bytes.length);
       rowBuffer.put(bytes);
     }
